refactor(runtime): decouple from proxy core
