+++ Sources/ProxyCore/DiscoveryClient.swift
@@ -42,7 +42,7 @@ package struct DiscoveryClient: DependencyClient {
 
     package static func live(
         defaultFileURL: @escaping @Sendable () -> URL = {
-            Discovery.defaultFileURL
+            ProxyFilesystemLocations.discoveryFileURL()
         },
         loadRecord: @escaping @Sendable (_ url: URL) throws -> DiscoveryRecord = {
             try Discovery.loadRecord(from: $0)
