+++ Sources/XcodeMCPRuntime/Bridge/UpstreamHealthManager.swift
@@ -1,7 +1,6 @@
 import Foundation
 import NIO
 import NIOConcurrencyHelpers
-import ProxyCore
 
 package final class UpstreamHealthManager: Sendable {
     package struct ProbeRequest: Sendable {
