+++ Sources/XcodeMCPRuntime/MCPProtocolVersion.swift
@@ -1,6 +1,12 @@
-import ProxyCore
-
 public enum MCP {}
+
+package enum MCPProtocolVersion {
+    package static let current = "2025-06-18"
+
+    package static func isSupported(_ version: String) -> Bool {
+        version == current
+    }
+}
 
 extension MCP {
     public enum ProtocolVersion {
